jaalPlayerModal CSS: no height, no overflow
It seems that removing the "height" and "overflow" rules in
exercise-player.css: #jaalPlayerModal allows the modal dialog of the
Player to be shown in an actual A+ LMS correctly: no scrollbar is
needed in the modal, but the window scrollbar works with the modal.
It is the Bootstrap library the does the trick in student's Exercise view.
Scrollbars in modal are unfortunately required in teacher's Inspect
Submission view, because there the grader feedback including the JSAV Player
data is shown directly on the page without a Bootstrap modal.

This problem has been added to the new TODO.txt.