Testbench: rename exercise html files
renamed:    DijkstraPE-research.html -> DijkstraPE-research+player.html
renamed:    DijkstraPE-research-v2.html -> DijkstraPE-research-v2+player.html

Both HTML files differ from the ones in the testbench of the *Recorder*.
Essentially, they have the code that includes the Recorder. We want to
prevent a situation where someone copies a newer version of the exercise
from the Recorder's testbench to the Player's testbench and overwrites the
Player-embedding code.